+++ gevent/pool.py
@@ -197,7 +197,17 @@ class GroupMappingMixin(object):
         return greenlet
 
     def apply(self, func, args=None, kwds=None):
-        """Equivalent of the apply() builtin function. It blocks till the result is ready."""
+        """
+        Rough quivalent of the :func:`apply()` builtin function blocking until
+        the result is ready and returning it.
+
+        The ``func`` will *usually*, but not *always*, be run in a way
+        that allows the current greenlet to switch out (for example,
+        in a new greenlet or thread, depending on implementation). But
+        if the current greenlet or thread is already one that was
+        spawned by this pool, the pool may choose to immediately run
+        the `func` synchronously.
+        """
         if args is None:
             args = ()
         if kwds is None: